Added spell upgrades/ improved spell weapon system
I changed the player weapon so that all spells will be called from the attack system. By doing this we can now get any spell out of order. I’ve also added a function to give the player a spell of a specific type. Now if the player decides to go to the fire level first or the water level first, they'll get their respective spells in order. I've also implemented an upgrade system, while you don't have to have the spells in the same slot as the first set of spells, they do need to be parallel to each other, primaries to secondaries. This unfortunately broke the spell UI display it does not display the correct type of spell when held I will work with Chris to help fix this later