Working on the win menu feedback
So I found out why coroutines are so annoying sometimes. Turns out they operate off of Time.timeScale, which we set to 0 when the game is paused. Meaning the coroutines will never start because time is frozen. very sad. fixed by using WaitForSecondsRealtime. Wasted so much time on this one :(