Push I wasn't able to make last night read description
Made the win screen look and feel better
Had to touch a few scripts for this one
started working on more animations and sounds for the enemies but I'll put that aside, I have other things that are higher priority to do.
My ISP is awful
Also removed some things from the project that we weren't using at all.